# docs/swagger.yaml
basePath: /
definitions:
...
  db.Key:
    properties:
      id:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
  db.Terminal:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      serial_number:
        type: string
    type: object
...
  db.UpdateKey:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  db.UpdateTerminal:
    properties:
      address:
        type: string
      name:
        type: string
      serial_number:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: RESTful API server for MIFARE transactions authorization
  title: MIFARE API
  version: "0.1"
paths:
...
  /keys:
    get:
      consumes:
      - application/json
      description: Возвращает список всех ключей
      produces:
      - application/json
      responses:
        "200":
          description: Список ключей
          schema:
            items:
              $ref: '#/definitions/db.Key'
            type: array
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение всех ключей
      tags:
      - keys
    post:
      consumes:
      - application/json
      description: Добавляет новый ключ доступа
      parameters:
      - description: Данные ключа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.Key'
      produces:
      - application/json
      responses:
        "200":
          description: ID ключа
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание ключа
      tags:
      - keys
  /keys/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет ключ по ID
      parameters:
      - description: ID ключа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешно удалено
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление ключа
      tags:
      - keys
    get:
      consumes:
      - application/json
      description: Возвращает данные ключа по идентификатору
      parameters:
      - description: ID ключа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные ключа
          schema:
            $ref: '#/definitions/db.Key'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение ключа по ID
      tags:
      - keys
    put:
      consumes:
      - application/json
      description: Обновляет данные ключа
      parameters:
      - description: ID ключа
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.UpdateKey'
      produces:
      - application/json
      responses:
        "204":
          description: Успешно обновлено
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление ключа
      tags:
      - keys
  /terminals:
    get:
      consumes:
      - application/json
      description: Возвращает список всех терминалов
      produces:
      - application/json
      responses:
        "200":
          description: Список терминалов
          schema:
            items:
              $ref: '#/definitions/db.Terminal'
            type: array
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение всех терминалов
      tags:
      - terminals
    post:
      consumes:
      - application/json
      description: Добавляет новый терминал
      parameters:
      - description: Данные терминала
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.Terminal'
      produces:
      - application/json
      responses:
        "200":
          description: ID терминала
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание терминала
      tags:
      - terminals
  /terminals/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет терминал по ID
      parameters:
      - description: ID терминала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешно удалено
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление терминала
      tags:
      - terminals
    get:
      consumes:
      - application/json
      description: Возвращает данные терминала по идентификатору
      parameters:
      - description: ID терминала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные терминала
          schema:
            $ref: '#/definitions/db.Terminal'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение терминала по ID
      tags:
      - terminals
    put:
      consumes:
      - application/json
      description: Обновляет данные терминала
      parameters:
      - description: ID терминала
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.UpdateTerminal'
      produces:
      - application/json
      responses:
        "204":
          description: Успешно обновлено
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление терминала
      tags:
      - terminals
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"